requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# AI/ML for recommendations
scikit-learn==1.3.2
numpy==1.26.2

# Scheduled jobs for reminders
APScheduler==3.10.4

# Email
Flask-Mail==0.9.1

# Development tools
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
faker==20.1.0

# Data handling
pandas==2.1.4

# Optional utilities
click==8.1.7
